refactor(reposicao): extrai SugestaoCard de AdminReposicaoNegociacaoParalela
Segundo corte do NegociacaoParalela. Extrai o render do card de sugestão
(renderSugestaoCard, ~117 linhas) pra SugestaoCard.tsx — componente
presentational que recebe a sugestão + o ranking-extra já resolvido +
callbacks. Os handlers de estado/mutação ficam na página (zero mudança
de data-flow): onMarcarVisualizada/onIrAoRanking/onMarcarEmAndamento/
onIgnorar/onFecharSemAcordo/onConverter.

Página: 1073 → 962 LoC. Limpa 7 imports órfãos que só o card usava (Eye,
MoreVertical, CardHeader, CardTitle, DropdownMenuItem, statusBadgeClass,
statusLabel) — pegos pelo noUnusedLocals do typecheck:strict (a página
já é strict).

SugestaoCard entra no strict (354).

Verificado: tsc 0, typecheck:strict 0, build OK, 423 testes passando.
